requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curlify==2.2.1
facebook_business==20.0.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-analytics-data==0.18.8
google-api-core==2.19.1
google-auth==2.30.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prettytable==3.10.0
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pydeck==0.9.1
pyecharts==2.0.6
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
streamlit-echarts==0.4.0
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
yarl==1.9.4
